Split out export
