+++ packages/ymir-core-cli/src/lib/plugin/export.ts
@@ -1,0 +1,49 @@
+import * as nodePath from 'path';
+
+import * as fs from '../config/helper/fs';
+import * as resolverLib from '../resolve/lib/resolver-operations/resolve-stack';
+import * as getPlugin from '../plugin/get-plugin';
+import * as getStack from '../stack/get';
+
+import { entriesToEnvFile } from '../dotfile';
+
+export async function _export(ymirPath: string, stackName: string) {
+  const stackSource = await getStack.stackSource(ymirPath, stackName);
+  const [defaultResolverErr, defaultResolver] =
+    getPlugin.defaultResolver(stackSource);
+
+  if (defaultResolverErr) {
+    console.error(defaultResolverErr.message, defaultResolverErr);
+    return;
+  }
+
+  const [resolverConfErr, resolverConf] = await getPlugin.configByStackSource(
+    ymirPath,
+    stackSource
+  );
+
+  if (resolverConfErr) {
+    console.error('Unable to create resolver config: ', resolverConfErr);
+    return;
+  }
+
+  const data = getPlugin.parseFiles(stackSource);
+
+  const resolved = await resolverLib.resolveStack(
+    ymirPath,
+    data,
+    defaultResolver,
+    resolverConf
+  );
+
+  const dotFileData = entriesToEnvFile(resolved);
+  const fileConf = data.stackConfig.FILE || data.defaultStackConfig.FILE;
+
+  if (!fileConf) {
+    console.error('No FILE config found in stack or default stack');
+    return;
+  }
+
+  const filePath = nodePath.join(ymirPath, '../', fileConf.path, fileConf.name);
+  return fs.writeFile(filePath, dotFileData);
+}
