+++ packages/ymir-core-cli/src/lib/plugin/index.ts
@@ -1,0 +1,9 @@
+import { _export } from './export';
+import * as get from './get-plugin';
+
+const plugin = {
+  export: _export,
+  get,
+};
+
+export default plugin;
